Inmobiliaria/requirements.txt
anyio==3.7.0
asgiref==3.7.2
CacheControl==0.13.1
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
Django==4.2.3
django-widget-tweaks==1.4.12
exceptiongroup==1.1.2
firebase==4.0.1
firebase-admin==6.2.0
google-api-core==2.11.1
google-api-python-client==2.91.0
google-auth==2.21.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
gunicorn==20.1.0
httplib2==0.22.0
idna==3.4
msgpack==1.0.5
proto-plus==1.22.3
protobuf==4.23.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.9
PyJWT==2.7.0
pyparsing==3.1.0
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
